Renamed EditorComponents and say-text to be less confusing
Renamed to EditorElements to not be confused with Vue components, since
there is now a MessagePreview component that, despite the previous name
of EditorComponents, cannot be referenced by state variables.
The reference "say-text" for the editor element has also been renamed to
"say-label" to better reflect its purpose.
