+++ tf2-chat-colouriser/app/utils/chat/editor-elements.ts
